refactor: Extract event construction from streaming.py into streaming_events.py
Move status parsing, citation handling, tool event classification, and
payload building into streaming_events.py (529 lines). streaming.py now
contains only stream orchestration (624 lines, down from 1,121).
All public names re-exported from streaming.py for backwards compatibility.
